First step in adding abstraction layer. Added input.rs and output.rs to handle the processing of the input file and processing of the output file.
